### 2022-08-05 #4641 @diehlpk
> @editorialbot invite @diehlpk as editor
>
> wave @diehlpk – would you be willing to edit this submission for JOSS?

Sure.
