### 2022-08-08 #4641 @cicirello
@diehlpk A couple years ago with a prior JOSS paper for a different library of mine, engnadeau was an amazingly good reviewer, providing just about the most useful constructive feedback I've ever seen. It looks like they changed their username--they are in the reviewer list as nnadeau (their username when they reviewed my prior paper), but current username appears to be engnadeau (definitely same person, as the new username is what currently shows on the issues and PRs they submitted on that other repo).

I'll look through reviewer list for some additional suggestions.
