### 2022-08-09 #4641 @pritchardn
Hello, I'm currently giving a review for another JOSS submission (hopefully finishing soon), which may slow me down a little, but this looks interesting and I will gladly review this.
