### 2022-08-09 #4663 @editorialbot
[REVIEW]: ρμ: A Java library of randomization enhancements and other math utilities
**Submitting author:** @cicirello (<a href="http://orcid.org/0000-0003-1072-8559">Vincent A Cicirello</a>)
**Repository:** https://github.com/cicirello/rho-mu
**Branch with paper.md** (empty if default branch):
**Version:** v2.4.0
