### 2022-08-24 #4663 @pritchardn
Hello, my review is basically complete and rather uneventful

This was a very smooth review to complete, I've run the examples, including the performance tests and made my own; everything went smoothly. The extensive README is really well put together and made reviewing very straightforward.
I have some comments, but nothing prevents my recommendation to accept this.

## Software Paper
Very well written, but given the performance tests which do exist in the codebase, it would be nice to see them mentioned in the paper. The methodology is itself interesting and would be useful to other library writers looking to verify their own performance claims. Given the efficiency of rho-mu's implementation (along with additional functionality of course) it is reasonable to include such a claim in the paper.
Another small thing, it would be nice to have the Github links be clickable in the generated pdf, but that is a minor nitpick.

## Documentation
Makes full use of standard Java documentation, the APIs look completely documented and re-iterates all required points on its homepage.

Happy to elaborate further on any points, but this looks like a really clean submission to me.
