### 2022-08-24 #4663 @cicirello
@pritchardn Thank you for the feedback. I'll revise to mention the performance tests in the paper.

The GitHub links do appear to be clickable for me, although it looks like JOSS's markdown processor doesn't change their font color to make it clear that is the case. It also appears to depend on how the pdf is viewed. When using the "view article proof on github" link earlier in this thread, it seems all links are disabled (citation links and urls). But when downloading the pdf and opening locally with either Acrobat or a web browser, the links work (even though the color is same as surrounding text). One of the 2 URLs is broken in the proof because it spans multiple lines, and the line number is being interpreted as part of the URL, but that should resolve itself without the line numbers.

I'm guessing that GitHub *might* be disabling links in PDFs when viewed directly from GitHub. I know they sanitize SVG images to disable various things, like links, and embedded javascript, etc. So maybe they do the same thing with PDFs.
