+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 
 # Summary
 
-The $\rho\mu$ library is a Java library of Randomization enHancements and Other Math Utilities. It originated as a core dependency for our other open source research software libraries, such as JavaPermutationTools [@Cicirello2018b] and Chips-n-Salsa [@Cicirello2020], providing these libraries with efficient implementations of randomization related algorithms. Chips-n-Salsa, for example, is a library of stochastic local search and evolutionary algorithms, which requires fast generation of random indexes into arrays and other structures, as well as fast generation of non-uniform random numbers necessary to implement Gaussian mutation, Cauchy mutation, etc. The $\rho\mu$ library includes implementations of efficient algorithms for randomly sampling indexes into arrays and other sequential structures, randomly sampling pairs and triples of distinct indexes, randomly sampling $k$ indexes, among others. It also includes efficient implementations of random number generation from distributions other than uniform, such as Gaussian, Cauchy, and Binomial. Additionally, it includes implementations of math functions that are required by the randomization utilities. The $\rho\mu$ documentation is hosted on the web (https://rho-mu.cicirello.org/), and source code on GitHub (https://github.com/cicirello/rho-mu), which also includes a directory of example programs illustrating library usage.
+The $\rho\mu$ library is a Java library of Randomization enHancements and Other Math Utilities. It originated as a core dependency for our other open source research software libraries, such as JavaPermutationTools [@Cicirello2018b] and Chips-n-Salsa [@Cicirello2020], providing these libraries with efficient implementations of randomization related algorithms. Chips-n-Salsa, for example, is a library of stochastic local search and evolutionary algorithms, which requires fast generation of random indexes into arrays and other structures, as well as fast generation of non-uniform random numbers necessary to implement Gaussian mutation, Cauchy mutation, etc. The $\rho\mu$ library includes implementations of efficient algorithms for randomly sampling indexes into arrays and other sequential structures, randomly sampling pairs and triples of distinct indexes, randomly sampling $k$ indexes, among others. It also includes efficient implementations of random number generation from distributions other than uniform, such as Gaussian, Cauchy, and Binomial. Additionally, it includes implementations of math functions that are required by the randomization utilities. The $\rho\mu$ documentation is hosted on the web ([https://rho-mu.cicirello.org/](https://rho-mu.cicirello.org/)), and source code on GitHub ([https://github.com/cicirello/rho-mu](https://github.com/cicirello/rho-mu)), which also includes a directory of example programs illustrating library usage.
 
 # Functionality
 
