### 2022-08-24 #4663 @cicirello
@pritchardn I was able to make the URLs look clearly that they are links. URL by itself = clickable link but regular text font color; but using markdown link syntax `[]()` fixes the color to make it clear it is a link. The "view article proof on github" still appears to disable all links, but otherwise works when viewing PDF locally.
