+++ paper/paper.md
@@ -27,10 +27,10 @@ The $\rho\mu$ library is a Java library of Randomization enHancements and Other 
 
 The randomization enhancements provided by $\rho\mu$ include:
 
-* Faster generation of random integers subject to a bound or bound and origin, utilizing the algorithm of @Lemire2019, as well as faster generation of streams of such bounded integers. A sample program demonstrates the speed advantage over the Java API, where on average $\rho\mu$ used 53.3% less CPU time than Java's `RandomGenerator.nextInt(bound)`.
+* Faster generation of random integers subject to a bound or bound and origin, using the algorithm of @Lemire2019, as well as faster generation of streams of such bounded integers. A sample program demonstrates the speed advantage over the Java API, where on average $\rho\mu$ used 53.3% less CPU time than Java's `RandomGenerator.nextInt(bound)`.
 * Faster generation of Gaussian distributed random doubles, with a Java port of the GNU Scientific Library's C implementation [@Voss2014] of the Ziggurat algorithm [@Marsaglia2000; @Leong2005].
 * Additional distributions available beyond what is supported by the Java API's `RandomGenerator` classes, such as Binomial and Cauchy random variables.
-* Ultrafast, but biased, `nextBiasedInt` methods that exclude rejection sampling to trade-off uniformity for speed, based on the approach of @Lemire2019, as well as streams of such biased integers. A sample program demonstrates the substantial speed advantage, where on average $\rho\mu$'s `nextBiasedInt(bound)` used 99.2% less CPU time than Java's `RandomGenerator.nextInt(bound)`, offering a massive speed boost for cases where strict uniformity is not required.
+* Ultrafast, but biased, `nextBiasedInt` methods that exclude rejection sampling to trade-off uniformity for speed, based on @Lemire2019, as well as streams of such biased integers. A sample program shows the substantial speed advantage, where on average $\rho\mu$'s `nextBiasedInt(bound)` used 99.2% less CPU time than Java's `RandomGenerator.nextInt(bound)`, offering a massive speed boost for cases where strict uniformity is not required.
 * Methods for generating random pairs of integers without replacement, and random triples of integers without replacement.
 * Methods for generating random samples of $k$ integers without replacement from a range of $n$ integers, including three alternative algorithms, reservoir sampling [@Vitter1985], pool sampling [@Ernvall1982], and insertion sampling [@Cicirello2022], as well as a method that chooses among these based on $n$ and $k$.
 * Methods to generate streams of random numbers from binomial, Cauchy, exponential, and Gaussian distributions.
