### 2022-08-24 78e572f
JOSS paper revision: include performance results (#93)
