### 2022-08-26 #4663 @xtruan
My review is complete. This package was a pleasure to review, I have no suggestions for improvement in the code, documentation, or paper. A couple of additional notes I made during the review:
- Main library built as expected from `main` on Windows 10 with JDK 18 + Maven 3.8.6
- Happy to see automated tests integrated as part of the Maven build process
- Separate examples library built as expected from `main` on Windows 10 with JDK 18 + Maven 3.8.6
  - 3 Javadoc generation warnings of the form: `warning: use of default constructor, which does not provide a comment` (this is exceedingly minor IMO)
- All examples run as expected
- Was able to import package version `2.3.2` from Maven Central into my own test project
- Documentation is fantastic and very complete
- Paper is excellent and meets all requirements
  - Verified the new additions resulting from @pritchardn's review, all look good
