### 2022-08-26 704e69b
added private const to example programs
