### 2022-08-26 #4663 @diehlpk
@cicirello

Some minor comments from the editorial perspective:

- [ ] Please add the country to the affiliation
- [ ] You have some statements like `used 53.3% less CPU` in your paper. I think here it would be good to mention the CPU you used for this test.
