+++ paper/paper.md
@@ -35,7 +35,7 @@ The randomization enhancements provided by $\rho\mu$ include:
 * Methods for generating random samples of $k$ integers without replacement from a range of $n$ integers, including three alternative algorithms, reservoir sampling [@Vitter1985], pool sampling [@Ernvall1982], and insertion sampling [@Cicirello2022], as well as a method that chooses among these based on $n$ and $k$.
 * Streams from binomial, Cauchy, exponential, and Gaussian distributions.
 
-The GitHub repository includes the data and t-Test results demonstrating the speed enhancements at extremely statistically significant levels, using Windows 10 on an AMD A10-5700 3.4GHz CPU with 8GB DDR3 RAM. 
+Comparisons used Windows 10 on an AMD A10-5700 3.4GHz CPU with 8GB DDR3 RAM. The GitHub repository includes the data and t-Test results demonstrating the speed enhancements at extremely statistically significant levels. 
 
 # Architecture
 
