### 2022-08-26 #4663 @cicirello
@diehlpk I added the missing country, and I added CPU/OS details. To avoid redundancy, I put the latter at start of the short paragraph that follows the bulleted list of functionality, rather than in the two bullets containing results. Also, in the repository itself, in the directory of example programs that also contains the output from my runs, I added a simple text file with the CPU/OS details as well. This way someone who doesn't read the JOSS paper will have those details as well.
