### 2022-08-26 #4663 @diehlpk
@cicirello

Please do the following:

- [ ] Generate a new release including all changes, you made during the review.
- [ ] Upload this release to Zenodo and post the DOI here. Note that the title and authors have to match with the paper.
