### 2022-08-26 #4663 @cicirello
@diehlpk Release created and archived with Zenodo. Here are the details:

Version: v2.4.0
Zenodo DOI: https://doi.org/10.5281/zenodo.7026452
